+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-14T07:36:34.304Z",
+    "2026-04-14T08:16:52.325Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {